backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
google-generativeai
pydantic-settings
python-dotenv
alembic
psycopg2-binary
